requirements.txt
hydra-core
matplotlib
pytorch-sidu
torch
torchvision
transformers
timm
detectors
gdown
pytorch_lightning
wandb
omegaconf
flatdict